refactor helper file
