+++ controllers/users.js
@@ -1,12 +1,13 @@
 const bcrypt = require("bcrypt");
 const usersRouter = require("express").Router();
-const Blog = require("../models/user");
+const User = require("../models/user");
 
 usersRouter.post("/", async (request, response) => {
     const { username, name, password } = request.body;
 
     const saltRounds = 10;
     const passwordHash = await bcrypt.hash(password, saltRounds);
+    console.log(passwordHash);
 
     const user = new User({
         username,
