+++ models/blog.js
@@ -2,6 +2,7 @@ const mongoose = require("mongoose");
 const blogSchema = new mongoose.Schema({
     title: String,
     author: String,
+    user: { type: mongoose.Schema.Types.ObjectId, ref: "User" },
     url: String,
     likes: Number,
 });
